🔎⚗️ ↝ Modifying a 3js package for generating worlds on the client
We've got some 3d planets being rendered in javascript, these are based on the initial C# algorithms in https://github.com/signal-k/starsailors/issues/4, https://github.com/Signal-K/starsailors-boilerplate/issues/1

We've got some plans to generate nfts from this: https://github.com/Signal-K/sytizen/issues/18. The models here are based on the C# model, we'll have an API call setup later so that they're both equivalent to each other directly.

Next step is setting up some image exports and saving user data, providing a way for users to export the 3JS graphic to their Supabase storage row. That will be commenced later on tonight.

All the fields will be set to be callable via an API & we will compare the output of different Deepnote graphs generated from `lightkurve` to provide further adjustments to either end.
One issue is that on the frontend for the generator, some of the components aren't selectable. This is because the original example components were built for a much older version of React>Next. Since these will be callable via a custom API, and the editor will be changed to be a form, that won't be an issue.

Docs available at the central issues on signal-k org

https://github.com/Signal-K/marketplace/issues/19
https://github.com/Signal-K/marketplace/issues/3
https://github.com/Signal-K/marketplace/issues/7
https://github.com/Signal-K/polygon/issues/26 -> These will be added to posts on the image/comment sharing part of the site (and later on Lens -> team members can refer to the roadmap) so that will be routed internally.