🛼🎨 ↝ New screenshot feature for the Planet Editor
Changelog:
Moved the gallery components into the offchain account section.
Added new feature where users can take a screenshot of the planet they've created

Currently, the screenshot is saved as an image URI string (see bottom of commit for e.g.), which is a problem as it is then uploaded to Supabase as a URI. What we need to do is convert the URI to an image inside the app (without saving it as a file on the app (as in a png file being saved next to the jsx file). This could be done before it's uploaded to Supabase, or when it's retrieved...

https://github.com/signal-k/sytizen/issues/18
https://github.com/signal-k/sytizen/issues/16