🤼‍♂️🎅🏻 ↝ Adding context functionality to client from proposals smart contract
This is the first step of our new attempt at fixing https://github.com/Signal-K/polygon/issues/25

I've started merging the CSS into a new `styles/Proposals` directory in the next client, converted the calculater functions to typescript, and adding the context as a wrapper around the rest of the app (inside Thirdweb -> Moralis [providers]).

We'll take the `<Sidebar />` component from https://github.com/Signal-K/sytizen/tree/main/client and have that be a persistent component on all pages INSIDE the `pages/Proposals` directory. We can remove the usage of `react-router-dom` by just putting a page inside that directory, and then creating a `[...id].tsx` file for the proposals dynamic routes, thus making it Next compatible. The context provider seems to be working fine. So we'll see how this goes.